Strenghten target/toolchain checks
* Target and toolchain variables are now window-scoped, in fact
many windows may be simultaneously be open and there is no reason that
they all have the same target/toolchain

* Target name is verified by comparing to `mbed target -S` output. This
check is weak as it can be produce wrong results (matching "A" is almost
always true for example). We need to search for the whole word, not the
pattern.

* Similarly, toolchain name is checked, it obviously must match ARM,
GCC_ARM or IAR.